requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
factory_boy==3.3.1
Faker==33.0.0
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
